+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Bump `suitable` to [0.6.0](https://github.com/clojure-emacs/clj-suitable/blob/v0.6.0/CHANGELOG.md#060-2023-11-05).
 
+## Bugs fixed
+
+* Make `"inspect-refresh"` op work again.
+
 ## 0.43.0 (2023-11-04)
 
 ### New features